IndiePantry_Backend/requirements.txt
appdirs==1.4.3
Django==1.9.12
djangorestframework==3.6.2
mysqlclient==1.3.10
packaging==16.8
PyMySQL==0.7.11
pyparsing==2.2.0
six==1.10.0
